Improve constraints test for being in a list
When the object is added to a hasMany relationship of another
domain object all objects will be validated again. If the test
are written in a way that they assume they are the new object
being added (and not an existing one) the constraints tests fail.

This adds a simple check to the tests to ensure that a already
saved object is not re-validated.

Of course this results in wrong values if someone changes the
values manually. We should protect that in a different way if
necessary.