Initial Implementation of Aspect for Job System
There is one Scheduler class which works as an Aspect for the
Job's execute method. It needs to be annoted with the JobExecution
annotation. The idea is to automatically inject this annotation
through an AST Transformation, so that the Developer does not
have to care about it.

The Scheduler currently intercepts before and after method execution
and ensures that some pre- and post-conditions are met. If not it
throws a RuntimeException. That should probably be improved to throw
a specific Exception inheriting RuntimeException. As well before
the exception is thrown a log message is written.

The Aspect adds ProcessingStepUpdates to the ProcessingStep of the
executed Job.

Currently there is one short integration test ensuring that the
basic behavior is working correctly. This has to be extended to
also test for Exception conditions as well as EndStateAwareJobs.

Also the Scheduler needs to be extended to store input and output
parameters and trigger the next Job in the execution, start the
error handling for the JobExecutionPlan and recognize an end state
and start the notification process for that.