Implementation of Job Error Handling in the Aspect
The after throwing aspect takes care of creating a Failure Update and
store an error message together with the update. For that a new domain
class ProcessingError has been added.

What is still missing is a service to log the stacktrace to a log file
and to generate a unique hash from it which will also go into
ProcessingError.

Unfortunately the after throwing advice is not able to "eat" the
exception that means that the service which later on triggers the Job
execution has to take care of wrapping the execution in a try/catch
block to ensure that the exception do not propagate too far.

Triggering the error handling is still missing in the advice as
it is in general not yet implemented.