AST Transformation for adding JobExecution annotation
This AST Transformation adds the JobExecution annotation to each
execute method of a class implementing the Job interface.
By using an AST Transformation the developer of a Job does not
have to care about such implementation details.

To trigger the transformation first the ant script has to be
executed to compile the jar for the transformation. This should
be changed into a compile event, so that nobody has to care about
using an ant script.