SchedulerService to start new Jobs
The Scheduler invokes the SchedulerService after a Job ended to
trigger the execution of the next Job. The Service decides whether
there is another JobDefinition and if there is one a new
ProcessingStep is created and added to a queue of to be executed
ProcessingSteps.

The actual starting of a Job is done through the schedule method.
This is currently not yet invoked automatically, but this will
change in future. The schedule method decides whether to start
a new Job and creates a Job for the ProcessingStep and executes
it and keeps track of it. After the Job finished (either success-
fully or with failure) the Job is removed from the internal List.

If the Job has been the last of a Process, the Process is set to
finished.

An integration test is added to ensure the basic functionality is
working correctly (no more Jobs and one more Job).