Split Parameter into ParameterType and Value
The ParameterType is a generic description of a Parameter including
a name combined with a JobDefinition. This makes the Parameters
specific to the JobDefinition and allows a better mapping between
the output and input parameters of subsequent Jobs.

As well we can introduce a uniqueness constraint on the name of
the Parameter.