Service for storing Exceptions to files
This Service stores Exceptions to files which occurred in Jobs.
It wraps the stacktrace of occurred Exceptions along with
the Job's class name and timestamp into xml and stores this
information to a file named with the md5 sum of the Exception
together with the Job's name. This way it is ensured that each
different Exception of each distinct Job has its own file.
When the same Exception occurs more than once for the same
Job only the timestamp is added to the xml as the other information
stays the same.