Create a Process through a StartJob
To identify the StartJob a new StartJobDefinition is introduced
which inherits JobDefinition and is linked to the JobExecutionPlan.
This means that the JobExecutionPlan does no longer store the
name of the startJobBean.

The SchedulerService can create the Process and add a ProcessingStep
for the JobExecutionPlan's first JobDefinition to its queue.

Creating and setting the Parameters is not yet tested.