The jobs to initially load meta data
Each of these jobs simply calls a method
of the MetaDataService which then processes
a distinct operation on the given run.
This run is different for each Process, so
that after all Processes are finished all
runs in the database shall be processed.
The start job fetched all runs and starts
a Process on each.