Add new classes for having process parameters
The ProcessParameter is a new parameter functionality
enabling to have parameters unchanged valid for a
complete process. These new parameters are this way
some kind of static for a particular process
simplifying parameters to be set by the start job,
used by several jobs, but not changed at all.
The to the ProcessParameter associated
ProcessParameterType is needed as it maps to the
JobExecutionPlan instead of a JobDefinition
(as ParameterType does) and as it is simpler.