Split out view specific JavaScript into an own file
All JavaScript goes into otp.js which uses a namespace OTP. An object
of this namespace is added to the jQuery object as "otp". So the
methods can be accessed through $.otp.

The OTP object takes the ${request.contextPath} as a ctor parameter
and the object is created in main.gsp, ensuring that the ctor is always
set.