The Notification Framework
This framework is able to send out messages either by mail or xmpp
to users configured as a group or implicit through an AclSid.

A Notification can be configured as a combination of a Trigger (e.g.
a JobExecutionPlan instance), the NotificationType (e.g. a Process
succeeded), the Medium (Mail or Jabber), a template for the message
and the Recipients for the Notification.

The Notifications can be triggered through application events.
Therefore a specific NotificationEvent is added and a specific
NotificationListener to wait for such events. All processing is
done in an own Thread, so that it does not block the normal job
processing thread.