Only one ProcessParameter per Process
Change the possible number of ProcessParameter
per Process from many to just one. It is agreed
that one ProcessParameter is sufficient. This
simplifies the application. The class
ProcessParameterType is no more needed and
all methods assigning the ProcessParameter
are simpler as well because the ProcessParameter
can be directly assigned now.
The relevant tests are also adapted.