Add possibility to limit number of processes
Now it is possible to limit the number of processes
for a particular JobExecutionPlan. The default value
of the specific variable is -1 indicating that there
is no limit. If a limit is senseful a positive value
should be set.
The method isNewProcessAllowed can be easily called
by StartJobs indicating whether it is allowed to create
a new process.