Support Clean Shutdown of OTP
For a clean shutdown a controller is added which allows to schedule
a shutdown. While the shutdown is active no new jobs will be started,
but new ProcessingSteps can be scheduled.

New Processes cannot be scheduled while the application is scheduled
for shutdown. For that the createProcess method in SchedulerService
has been adjusted to provide a boolean return value which is false
during shutdown process.

The ShutdownService and Controller provide information about which
Processes are still running and whether these can be suspended.
The administrator can finally stop the applicationContext through
a button in the web interface. This will trigger the application
to go down and ShutdownService is invoked to suspend all restartable
Jobs. If there are non-restartable Jobs it will log a warning.

All Shutdowns are logged in the database including the who, when and
why. If a clean shutdown succeeds it will get the proper date and
in case the shutdown gets canceled it's logged as well.