Refetch the ProcessingStep when creating a Job
Rational: the ProcessingStep might have been created in a different
session than the one bound to the Thread which tries to access the
ProcessingStep causing a Hibernate Exception. By refetching this
problem does not occur.

While this should be a safe change it has to be tested in a highly
concurrent situation nevertheless.