Resume Jobs after Server Shutdown/Startup
This is the clean Shutdown case: the scheduler is stopped till
the method startup() is invoked which is done by the Bootstrap.

The startup() method inspects all unfinished ProcessingSteps and
tries to add them to the queue again. If the startup() method
detects a ProcessingStep which indicates an unclean shutdown,
all changes are rolled back and the system has to be fixed manually
before the scheduler can be started.