Restarting of failed ProcessingSteps
Failed ProcessingSteps can be restarted. For that a new status
RESTARTED is added and the ProcessingStep is added back to the
queue. This behaves the same as a created status.

Restarting is supported in crash recovery, from Process view (for
the last failed one) and for ProcessingSteps view.

To support this the datatables are extended by an "actions" collumn
which renders a link to restart the ProcessingStep.