Add view for all failed Processes of a given Plan
For this the Workflow overview table is extended by another
column showing the numbers of failed processes for each plan.
It is a link leading to a view just showing the failed processes.
The view is shared with the normal view but a parameter is added.
As well parameters are added to the service methods to perform
a restriction for the last state of a Process.