Show logged stacktrace from ProcessingStep view
In order to show the stacktrace the Scheduler also stores the
stacktrace identifier into the ProcessingError, which already
had a field for this piece of information.

ErrorLogService is extended to extract the actual stacktrace
from a stacktrace identifier and ProcessService provides a
secure method to retrieve the error for a given ProcessingError.

The User Interface is just extended to render a link for the
error, if a stacktrace is present. The stacktrace itself is
opened in a new browser tab.