Enable/Disable a JobExecutionPlan
An admin user or a user having write access to a JobExecutionPlan
is able to enable/disable it from the Plan View page.

For this the view is extended with a message whether the Workflow
(Plan) is currently enabled or disabled. This is visible to all
users having access to the JobExecutionPlan.

For all users who are allowed to alter the state, a button is
added to enable/disable the plan. Unfortunately some HTML code
had to be duplicated to handle admin user case and ACL case.

Through JQuery an AJAX request is sent and based on the returned
result the button and text is updated.