Adjust to the needs, provide wanted API
The ExecutionServer now has the API required for
stabil usage. The API consists of the 3 public
methods on top of the classes body. These methods
just call the functionality excecuting the
command/script/job on the specified server.

The ConfigService is extended and retrieves now
all needed parameters from the Config. These
parameters are used by the ExecutionService.