Fix JobExecutionPlanService.getLastExecutedProcess
The method actually returned an incorrect result which was just
not found by the unit test but got uncovered by using to the
integration test.

A direct HQL query is now used to directly find the correct
Process by selecting the created updates for all first
processing steps and using the max process found by that.