Extend JobExecutionPlanDSL to define a watchdog for a job
The closure to create a JobDefinition is extended by:
watchdog(String beanName)

If this is present as the last command in the jobDefinition
closure a new JobDefinition is created with name:
__WatchdogFor__<jobName> using the specified bean.

Automatically an output parameter __pbsIds is added to the
JobDefinition <jobName> and mapped to the input parameter
__pbsIds of the watchdog JobDefinition.

The watchdog becomes the next JobDefinition of <jobName>,
which means the watchdog is inserted in between the normal
JobDefinitions. The user of the DSL does not have to care
about the fact that another additional job is created at all.