PbsJobs provide a list of Process Ids on PBS
A job implementing the PbsJobInterface needs to provide a list
of the Process Ids it started on the PBS.

The Scheduler takes care of creating an output parameter for
these Process Ids for the reserved parameter type __pbsIds.

Also the Scheduler verifies that the PbsJob provides these
Process Ids and raises an error if not present.