PbsJob provides the ID of Realm it is operating on
A PbsJob is operating on a Realm. The id of this Realm
is provided through a getter and stored by the Scheduler
as an output parameter. This allows to pass the same
realm to the Watchdog Job running after the PbsJob.

In the tests also a constant input parameter is created
for the PbsJob containing the same realm id.