Define pbsJobs in the JobExecutionPlan DSL
The pbsJob definition "inherits" the job definition
and extends it by a "realmId" which is added as
a reserved constant parameter called "__pbsRealm".

A pbsJob should be used together with a watchdog and
as soon as there is the generic watchdog it will no
longer be required to specify the watchdog, but it
will be created implicitly.