Working war files when deploying to Tomcat
Changes to make the war work when deployed to Tomcat and to
make grails run-war working.

To get the war working in the first place one class from the
ast directory needs to be copied to the src directory. This is
done by the _Events.groovy script by copying the file to the
right location on Compile Start and deleting it again at
Compile End. The class is used by both AST and otp, that's why
this hack is needed. If the ast lib which is added to the
compile path packaged into the war, the war is not starting up.

Furthermore for run-war the tomcat opts in BuildConfig.groovy
adjusted so that tomcat does not run out of memory and the
database config is adjusted for production use.

In Config.groovy there is the grails.serverURL which needs to
be adjusted for productive usage.