Production database properties are configurable
The connection properties of the production database are now set
via properties file. This is done to make them easily configurable.
Just add the connection details to your .otp.properties file.
It is still assumed that PostgreSQL is the only target
database server system.
If the dbCreate option is not set in the properties file it is set
to 'validate'. Validate ensures that Hibernate does not change
database schema but throws an exception during application's startup if
database schema and data objects differ somehow.