Refactoring of JobExecutionPlan DSL
All the specific closures are defined on class level as a
private static closure. The actual closures needed while
executing the DSL are created through currying.

This refactoring allows to share the code between e.g. the
startJobDefinition and a normal JobDefinition and makes it
easier to extend in future, e.g. for adding a validatingJob.