Introduction of a RestartedProcessingStep
When a ProcessingStep failed and should get restarted by the
SchedulerService a new RestartedProcessingStep is created which
is linked to the original failed ProcessingStep.

The RestartedProcessingStep takes over the next/previous link
in the ProcessingStep chain. That is the next link of the
failed ProcessingStep's predecessor is changed to point from
the failed one to the new RestartedProcessingStep.

This means the chain of ProcessingSteps is kept to point to the
latest chain of executed ProcessingSteps.

At the same time it's possible to go back to the list of
ProcessingSteps which got executed before.

This slightly changes the meaning of RESTARTED state. RESTARTED
is now a final state for a ProcessingStep indicating that a new
ProcessingStep has been created.