General data source section is now standard
The generic data source section is now standard for
production and test environments. Therefore these
specific environments are empty.

The test data source is configured independently to
work correctly with h2. Additionally, the catch block
of the config setting section is adjusted to use h2
as well.

The custom properties file is now set correctly
and tested whether the file exists.