ACL protection for individual and refactoring of Individual Controller
Move business logic of Individual Controller into the
IndividualService and at that time add ACL protection to the
methods. The protection is ensured on the Project level or admin.

Some of the Database access is moved into the domain classes
instead of having it in the service level. This is mostly about
data belonging to the Identifier, so to say making usage more
easy.

At the moment the list of Individuals is still unprotected and
the start of igv needs to be seriously reworked. The code in
it's current state cannot be protected and is extremely crashy.

REVIEW: 32