Inject an ExpandoMetaClass into JobExecutionPlanDSL
When running from script the closures inside the DSL use a
ClosureMetaClass which cannot have new properties injected.
When accessing the DSL from the running system an
ExpandoMetaClass is used which allows to add further closures.

With this change we create an own ExpandoMetaClass for each of
the closures and by that ensure that it is always possible to
inject, no matter from where the DSL is called.

REVIEW: 38