Ajax-ified Run list
An ajax version for the list of Runs is added. In order to have
all dtatables behave the same the code is split out into a
generic part which is the same for each view.

The Runs are retrieved through the RunService providing
pagination, sorting and filtering. Though ACL is not yet
available.

REVIEW: 40