New tag: editorSwitch
An otp tag library is introduced with currently one tag called
editorSwitch. This tag provides a read-only text field with a
button to start editing. If pressed it switches to text edit
with a button to submit and to cancel.

Updates are sent through AJAX to a specified resource with
feedback provided in the global notification area.