Support for editing MetaDataEntries from DataFile Controller
The MetaDataService provides an ACL aware method to retrieve
a MetaDataEntry based on it's ID and provides a way to update
the value by updating the Changelog. In case it fails two
specific exceptions can be thrown.

The DataFileController is extended to provide an action to
update a MetaDataEntry and uses the new editor-switch
component to render a useful editor component.