Rework main home page of OTP
The auto-generated controllers and plugins are removed.
Instead an introduction text is displayed which is taken from
messages.properties. This text can probably be improved.

For administration users a few admin links are displayed. There
the CSS needs some adjustments, the li elements have a
missplaced bullet.

In addition the main layout is extended by a home button.

REVIEW: 48