Drop completely scaffolded controllers
None of these controllers is linked from any other controller or
view. Which means all of them are orphaned and cannot be
accessed. For most of the data which would be represented by
these controllers, the data is integrated in other views where
it makes more sense to show it. E.g. MetaDataEntries are
included in the DataFilesController.