Adding a ChangeLog view
The ChangeLog view is implemented in the OTP tag library. It
provides a button which will perform an AJAX query to retrieve
the ChangeLog data and renders it in a jQuery UI dialog.

As an example the ChangeLog is added to the MetaDataEntries of
the DataFile controller.

The button is not yet styled. Some proper CSS should be added
for it.

REVIEW: 52