Domain class to reference another domain class from ChangeLog
A new ReferencedClass domain class is added which just holds
one field: className.

This class is meant to be used anywhere where a loose reference
to another domain class is required. This is added to ChangeLog
in order to have faster join based queries and smaller tables.

To be able to transit the existing data, the new field is added
as a nullable column and the old tableName allows to be blank.
In addition there is a migration grails script which moves the
existing data to the new format.

New entered ChangeLog entries will only generate data of the
format which would allow in a future release to drop the field
altogether from the database with an alter table SQL script.

REVIEW: 54