Refactoring of my* CSS classes
The myFoo classes mostly had the usecase of having a header
with a table underneath. This is no properly supported by
nested less definitions. To make proper use of that some gsp
files had to be mostly rewritten (yes html is not nice for
version control systems). Basically there was a re-grouping to
have a div with a nested header and table.

Css classes which are now unused in the codebase got completely
removed.