Improve ACL definition in MetaDataService
The ACL definition did not cover the case that the DataFile does
not yet have a project assigned which results in a crash.

Also the ACL definition on the update call are just wrong by
allowing access on read instead of write.

To properly verify an integration test is added.

BUG: OTP-6
REVIEW: 58