Adding a group for assigning ACL to users
The new Group can be created through the edit User view in a
dialog. A Group takes name and description and a set of
boolean values to indicate whether ACL should be created for
* Projects
* Sequence Centers
* Job System

In all cases there is a read or write permission. The idea is
that the creation of JobExecutionPlans, Projects and co
inspects all Groups and sets the ACL accordingly.

For each Group a Role is created with the same name as the
Group, but prefixed with "GROUP_", all uppercase and white
spaces replaced by underscores. The UserService is adjusted to
filter out these special roles when asking to list them.