Add security tests for the JobExecutionPlanService
Each protected method of the JobExecutionPlanService received a
test method to verify that ACL and role based filters work
correctly. That is it is verified that a user does not have
access, an admin has access and after setting the ACL the user
gains access, but other users are unaffected.