Adding a ROLE_OPERATOR
The operator is a user role for all tasks related to the operation
of the NGS pipeline. That is both the NGS data and the processes.
The idea is to have the operator everywhere were we currently
use admin privs but does not actually require an admin.

The operator is a user inheriting the privs of ROLE_USER and the
ROLE_ADMIN inherits from ROLE_OPERATOR. That way everything which
is allowed for an operator is also allowed for an admin.

BUG: OTP-12