Do not animate status icon for a RESTARTED ProcessingStep
For a restarted ProcessingStep a new ProcessingStep is created.
This one is animated for being running. The step which got
restarted will always be in the state restarted, so it should
not use the animated icon, but the normal red icon to indicate
that the step had failed before.

BUG: OTP-37
REVIEW: 66