Auto-refresh functionality for List Views
The Process Views can auto-refresh their data. Therefore a new
tag is introduced which renders a div with two links:
one for enable the feature, one for disable

Both are handled by a new RefreshController which just stores
into the session whether auto-refresh is enabled or disabled
and return the current state as a JSON string

Through JavaScript the shown state of the two links are swapped
and a boolean flag is kept to indicate whether it's currently
enabled.

This flag is used in a timer specified on the shared list view
implementation of the Process related views. The timer tries to
reload the data every ten seconds if auto-refresh is enabled.

BUG: OTP-38
REVIEW: 67