Logging system for Jobs
Each Job logs into an own log file in:

logs/process-id/processing-step-id.log

with process-id being the database Id of the Process the Job is
running for and processing-step-id the respective Id of the
ProcessingStep of the Job.

As the underlying logging infrastructure of log4j is difficult
to configure dynamically at runtime some changes are required.

The AST transformation injects a Log property into each Job as
__internalLog. In addition a private Log property as log is
injected, so that each Job can just use log like it is done in
any other class.

Unfortunatelly grails dynamically creates the log property if
it does not exist. This makes it impossible to just define the
log property in the abstract parent class, but we need to
inject it in the AST transformation.

When the Job is created by the SchedulerService the log
property is set to an instance of the JobLog class. This is an
implementation of the Log interface wrapping the log messages
into a JobLogMessage instance containing the original logged
message and the ProcessingStep of the Job. This JobLogMessage
is logged to the real Log, that is the injected __internalLog.

This ensures that the logging message ends up in the normal
log4j system. For logging the messages of the Jobs into the
files a custom JobAppender is used. This Appender recognizes
JobLogMessages and passes the original message to a FileAppender.
For each ProcessingStep a different FileAppender is used. When
a ProcessingStep finishes the JobAppender is informed to close
the FileAppender for the ProcessingStep.

To complicate the whole system log4j calls toString() on the
logged message before passing it to the Appenders. This means
that casting the message of the LoggingEvent to JobLogMessage
is not possible. To circumvent this problem the JobLogMessage's
toString() method serializes the object and returns a Base 64
encoded representation of the serialization. The JobAppender
decodes the message and deserializes the JobLogMessage.