Better visual identification of running system
A new tag is included in the otp taglib to render the environment
name. This name is taken from the otp.environment.name property.

If this property is not present the Grails environment name is
used. E.g. if run on production the name would be production.

This tag is used in the class definition of the main html tag of
the main layout. In addition the name is rendered into the
navigation area in case it is not production.

The main background of OTP is adjusted for the environments:
* production
* alpha
* beta

In all three cases the colors are adjusted to be more agressive
the more "dangerous" the environment is. That is development as
the moust "dangerous" environment is red.

BUG: OTP-45
REVIEW: 77