ACL aware notifications
The Notification Trigger optionally includes an ACL check. This
is performed by the NotificationListener. In case that an ACL
check is required and the Recipient of the Notification is a
SidRecipient the Acl for a domain object is retrieved and the
Recipient is only included iff read permission is granted to
the domain object.