Controller to modify the Notifications
The new NotificationController allows to modify the existing
Notifications. It is not possible to create new Notifications in this
controller - use the DSL instead.

Following functionality is supported:
* enable/disable notification
* Change the type (process started, failed, etc.)
* Change the medium (email, jabber, etc.)
* Change the trigger for what the notification should be sent
* Change the subject
* Change the message

All Notifications are shown in a datatable and editing anything of the
notification is done directly in place. E.g. a checkbox to enable the
notification will directly send an AJAX request to perform the action.

Instant feedback is provided in the AJAX responses always updating the
datatable to have the latest state after an operation succeeded or
failed.

All data going into the controller actions are validated by a Command
object and in case it has errors is rendered as JSON back to the JS which
uses it to show a proper error message.

The operations are actually all performed by a service which ensures ACL
protection and returns whether the operation was performed successfully.
As for this controller detailed information is not required it does not
provide an error message.