Use git branch name from environment variable if available
In Jenkins the name of the branch which is built is available through the
environment variable $GIT_BRANCH. The existing logic to extract the name
of the current branch is not working in Jenkins as Jenkins performs some
internal git magic.

If the environment variable is set it is used to include the git branch
name, if not available the git repository is queried, so that a local
build does not require the env variable to be set.

BUG: OTP-55