Introduction of FastqcProcessedFile
This object keep information about existing fastqc result file.
The domain classes from ngsqc package now have reference to this object
not directly to DataFile.

The service FastqcUploadService was split into part which does upload and
second part with produces statistics. The second part is not part of this commit.