Fix to the bug stalling application
In FastqcJob the jobId parameters where not properly exctracted and contained
the server name. This cause no problem when querying PBS in the proper realm but
while querying other realms "qstat" was taking many minutes. If 8 jobs where
stalled in waiting for PBS ansver the whole application was stalled.

Solution - proper extraction of the PBS job ids.