Enable user switching for Admin users
In the User Administration overview table each row for a user is extended
by a button to switch to this user. If an admin user clicks this button
he performs a "sudo -u" to this user and will operate in OTP as if he
were this user. This allows to e.g. easily test whether the ACL is setup
correctly.

When being switched to another user a link next to logout is added to
switch back to the primary account.