Backdoor allowing users to auto-login in DEVELOPMENT mode
A BackdoorFilter is added which is injected into the security filter
chain in development mode. This filter will set the Authentication in
the SpringSecurityContextHolder to contain an Authentication created by
this filter with a configured username and role user and admin as
Authorities. That means the developer does not have to log in.

As this is quite a backdoor there are multiple checks to ensure that the
filter will never ever be run in production mode:
1. The filter is only injected into the chain if we are in development
   mode
2. The filter itself checks that it is used in development mode
3. It's configured through config options otp.security.useBackdoor and
   otp.security.backdoorUser. If one of the two config options is not set
   the filter disables itself. Also if useBackdoor is false
4. The config options are only mapped from properties file in
   Config.groovy in an environment block for development

If one of the checks fails the filter will not allow access. What should
nevertheless be done is to remove the compiled class file from the war.

REVIEW: 97