Pass parameters to a restarted Job
So far a restarted Job did not get the required input parameters, now the
output parameters of the previous run Job in the execution plan are
passed to the restarted Job.

For normal situations this is sufficient, but in case it's the first job
of the execution plan it's possible that parameters get lost. This is
difficult to solve and at the moment a theoretical problem as we don't
have any execution plans which would run into this issue.

BUG: OTP-57
REVIEW: 103