Workflow to calculate Summary status of fastqc
After all seqTracks for a given run are processed with FastqcWorkflow
this workflow loops over results and calculate summary number.
Thie number is set in the Run object and is displayed in the main
view of the run (with sorting)

Current implementation if focused on semaphoring and infrastructure
and the algorithm to calculate actual number will be improved taking
into account different importance of fastqc modules.