Alignment Workflow
Alignment Workflow runs BWA alignment for one seqTrack.
Executing alignment (creation of the Sai files), pairing and sorting (bwa and samtools)
and removing duplicated (picard). The workflow assumes the tools: bwa, samtools and picard
are installed and directly called from the shell.

Each file created is book-keeped and file existence is validated in validation jobs.
Files created are book-keept in domain classes and corresponding services
are used to calculate files path in the filesystem.

Alignment workflow introduces Options Framework which allows to keep options
of data processing in a database.