Introduce dedicated switch user role
A new controller SwitchUserController is introduced which allows a User
with the ROLE_SWITCH_USER to switch to another user which is neither
Admin nor Operator. An admin user can still switch to any other User
through the UserAdministrationController.

For this the ROLE_SWITCH_USER is introduced which is automatically
passed to each Admin user and the user switching is now protected by
this role instead of role admin.

Unfortunately spring security does not provide an additional restriction
on the user one wants to switch to, only on the User who wants to switch.
This means any user with the given role is able to gain higher privileges
in case he knows the actual internally used URL. So to say we have here a
security by obscurity feature.

Though for the given use case this seems to be good enough given that the
target user group of this feature has no internal understanding of Spring
and would not know how to write the specific URL, also the URL is never
shown in the browser.

REVIEW: 118