New view as Overview over all sequences
The new Sequences view joins all data related to the sequencing track.
Therefore a database view is created (see scripts/sequences.sql) and
mapped to a new Grails domain class. This allows to easily search and
filter on the data referenced from the service. Instead of complex HQL
queries the search and filtering can be performed through readable
criteria.

In the web view a datatable is used to render the data and the filtering
is performed through a set of comboboxes with controlled vocabulary or
search fields. It is possible to combine all the options which results in
and-ing the filter and for multiple filters on the same type those are
or-ed but still and-ed with the other options.