Use proper classes to represent PlanInformation
The generated PlanInformation used to use a very simple Map which does
not provide any type guarantees or proper information of what it
contains.

This is now replaced by a few classes containing all the information and
gathering it together in a PlanInformation. A PlanInformation contains
all the information of all the Jobs and Parameters and the Jobs even
include information about a concrete executed ProcessingStep.

REVIEW: 135