Scheduler cannot be restarted if there is a RestartedJob
In the case that there is a ProcessingStep which is in restarted state
then there should be another RestartedProcessingStep being linked to the
ProcessingStep in restarted state.

The ProcessingStep in restarted state does not have a next ProcessingStep
which means that to the Scheduler it looks like the last one in the chain
which is not correct. The Scheduler incorrectly spotted an inconsistency
there and refused to startup.

This change checks whether there is a RestartedProcessingStep available in
case of a ProcessingStep in state restarted and the Scheduler will only
spot an inconsistency if such a RestartedProcessingStep is missing.