Better handling of contextPath in JavaScript
Instead of setting the contextPath manually in the main layout after the
JavaScript file has been processed by the browser, it is now written into
the meta information which is known to be processed before the JavaScript
file gets loader (because it is on top of it).

This allows to have the JavaScript file reading the contextPath itself
and no hardcoding is no longer required for i18n.

REVIEW: 152