Create new Users
A new view is added in UserAdministration to create new users. As an
experiment jQuery Validation plugin is used to validate the input fields
on the fly.

To create the User a new CommandObject is created which contains the
User relevant fields and List of Role ids and List of Group ids. If
valid it is passed to the service to create a User from it and add the
newly created user to Roles and Groups.

The saving of Users is adjusted to no longer encrypt the set password,
but replace it by an asterisk, this is needed as the password is taken
from LDAP.

REVIEW: 169