Improve checking of PBS Ids in ExecutionService
The logic to query one Realm for one pbsId is split out into a private
method. This is used by a new method to check a list of pbsIds on a
dedicated Realm instead of all Realms. This is useful if a Job knows that
all pbsIds are on one Realm.

In addition the method which queries all Realms is adjusted to not query
further Realms if it found one running. We are interested in the running
state, so if a pbs job is running further Realm will not give us better
information.