Introduce a PbsMonitorService
The PbsMonitorService queries the PBS in a time scheduled manner for all
PbsWatchDogs implementing MonitoringJob. Unfortunately specifying the
@Scheduled annotation on the PbsMonitorService's method is not working
due to limitations in Grails. It is therefore mapped in the
SchedulerService where the annotation is functioning.

The new service holds a Map of MonitoringJobs with the Pbs IDs they
registered. Whenever one of the pbs IDs finished on the PBS the service
notifies the MonitoringJob.

It is the task of the implementing MonitoringJob to know when all Jobs
finished.

Given that the MonitoringJob is not performing any tasks the execute
method of this Job is allowed to finish before the Job determined whether
it succeeded or failed. The doEndCheck Aspect in Scheduler is adjusted to
take care of these special kind of Jobs and does not perform the end
check for them.

Instead the logic is moved into the SchedulerService and once all jobs
on the PBS which are watched by the MonitoringJob have finished, the
MonitoringJob invokes the doEndCheck in the SchedulerService directly.

This is a fundamental change in how the Job System works. So far it was
not allowed for a Job to interact with the Scheduler directly. But it
needs to be changed as the execute method is blocking the thread pool and
there is no need to keep the threads running for days and doing actually
nothing. In some way the watchdog jobs can be considered as part of the
Scheduler so it seems to be just fine to change the behavior.